fix(react): use react in peerDependencies
Description
---

Register `react` and `react-dom` dependencies in `peerDependencies` instead of `dependencies`.

Motivation and Context
---

The `@tari-project/tari-extension-query-builder` is supposed to be used in external applications as well.
And they might use different `react` version from one used in this project.
If `react` versions differ, react loads two different instances of react runtime and this `QueryBuilder` component fails to load.


<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify