fix: serialize bigint
Description
---

Fixes bigint serialization by introducing a wrapper type. #22

Motivation and Context
---

Flows with U32 and other numeric types could not be saved, since it yielded an error:
"Do not know how to serialize a BigInt".

That is because JSON.stringify() does not support BigInt serialization. We will introduce a wrapper type.

How Has This Been Tested?
---

Locally - 
<img width="681" height="256" alt="Screenshot 2025-07-23 at 13 45 19" src="https://github.com/user-attachments/assets/c31534a3-50c3-4496-812f-145139a8f7f0" />

Serialized:
```json
{
        "values": {
          "value": {
            "success": true,
            "data": "22"
          }
        }
}
```

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a database, resync the chain -->